code/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
cfractions==2.4.1
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
cryptography==45.0.4
Django==5.1.6
django-ninja==1.3.0
django-ninja-simple-jwt==0.6.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
ground==9.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
locus==10.0.0
locust==2.32.10
locust-cloud==1.23.2
MarkupSafe==3.0.2
msgpack==1.1.1
pillow==11.1.0
platformdirs==4.3.8
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-engineio==4.12.2
python-socketio==5.13.0
pyzmq==27.0.0
reprit==0.9.0
requests==2.32.4
setuptools==80.9.0
shewchuk==6.10.0
simple-websocket==1.1.0
sqlparse==0.5.3
symba==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
